Polish console review nits: dedupe node helpers, SPA link, bounded log scan
Consolidated non-blocking review follow-ups from the two Console exec plans: the duplicated TaskNode/BranchNode edge-degree helpers move to a shared node-edges module, JobDAG's maxZoom is single-sourced, the queue-row Inspect anchor becomes a TanStack Link (no full page reload), and the structured-log heuristic is bounded to the head of the buffer plus short-circuited once the structured renderer locks in (no more O(n^2) rescans on streaming logs).
